Use a new module name
