+++ delegate-processors/src/main/java/com/palantir/delegate/processor/AnnotatedTypeMethod.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright 2023 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2024 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.palantir.delegate.processors;
+package com.palantir.delegate.processor;
 
-import com.palantir.delegate.processors.ImmutableAnnotatedTypeMethod.ImplementationBuildStage;
+import com.palantir.delegate.processor.ImmutableAnnotatedTypeMethod.ImplementationBuildStage;
 import java.util.List;
 import javax.lang.model.element.ExecutableElement;
 import org.immutables.value.Value;
