+++ delegate-processors/src/main/java/com/palantir/delegate/processor/Methods.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright 2021 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2024 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.palantir.delegate.processors;
+package com.palantir.delegate.processor;
 
 import com.google.auto.common.MoreElements;
 import com.google.auto.common.MoreTypes;
-import com.palantir.delegate.processors.DelegateProcessorStrategy.DelegateMethodArguments;
+import com.palantir.delegate.processor.DelegateProcessorStrategy.DelegateMethodArguments;
 import com.palantir.javapoet.AnnotationSpec;
 import com.palantir.javapoet.CodeBlock;
 import com.palantir.javapoet.MethodSpec;
