+++ versions.lock
@@ -8,7 +8,7 @@ com.google.guava:listenablefuture:9999.0-empty-to-avoid-conflict-with-guava (1 c
 com.google.j2objc:j2objc-annotations:3.0.0 (1 constraints: 150aeab4)
 com.palantir.goethe:goethe:0.14.0 (1 constraints: 37052f3b)
 com.palantir.javapoet:javapoet:0.5.0 (2 constraints: c7103ed1)
-com.squareup:javapoet:1.13.0 (1 constraints: f50b65f7)
+com.squareup:javapoet:1.13.0 (2 constraints: 2b113eee)
 org.checkerframework:checker-qual:3.43.0 (3 constraints: 2727263b)
 org.immutables:value:2.10.1 (1 constraints: 3605303b)
